refactor(api-aco): share Path with folder and flp entities
